requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
async-timeout==4.0.1
attrs==21.2.0
certifi==2021.10.8
charset-normalizer==2.0.9
cycler==0.11.0
fonttools==4.28.4
frozenlist==1.2.0
idna==3.3
kiwisolver==1.3.2
matplotlib==3.5.1
multidict==5.2.0
numpy==1.21.4
packaging==21.3
Pillow==8.4.0
pyparsing==3.0.6
python-dateutil==2.8.2
python-wootrade==0.1.9
requests==2.26.0
six==1.16.0
typing_extensions==4.0.1
urllib3==1.26.7
websockets==10.1
yarl==1.7.2
